More organisation of .py files.
