+++ src/qplot/windows/_widgets/treeWidgets.py
@@ -10,8 +10,20 @@ from qplot.datahandling import (
     )
 from .preview import (
     COLLAPSE_MINIMUM_RATIO,
-    DraggablePreviewImageLabel,
     PreviewTab,
+    )
+from .details_tables import (
+    CopyableTableWidget,
+    WrappedValueDelegate,  # noqa: F401 - re-exported for compatibility
+    format_value,
+    infoTree,
+    snapshot_parameters,
+    )
+from .run_list_items import (
+    EqualsAlignedDelegate,
+    MEASUREMENT_PREVIEW_SIZE,
+    RunPreviewCell,
+    SortableTreeWidgetItem,
     )
 from ._run_formatting import (  # noqa: F401
     format_complete_cell,
@@ -41,252 +53,6 @@ from os.path import isfile
 import numpy as np
 
 from datetime import datetime
-
-from .._shortcuts import standard_key_sequences
-
-COPY_SELECTION_SHORTCUTS = standard_key_sequences(QtGui.QKeySequence.StandardKey.Copy, ["Ctrl+C"])
-COPY_CELL_SHORTCUTS = [QtGui.QKeySequence("Ctrl+Shift+C")]
-MEASUREMENT_PREVIEW_SIZE = 22
-MEASUREMENT_PREVIEW_SPACING = 3
-
-
-def copy_action(label, shortcuts, slot, parent):
-    action = QtGui.QAction(label, parent)
-    action.setShortcuts(shortcuts)
-    action.setShortcutContext(QtCore.Qt.ShortcutContext.WidgetWithChildrenShortcut)
-    if hasattr(action, "setShortcutVisibleInContextMenu"):
-        action.setShortcutVisibleInContextMenu(True)
-    action.triggered.connect(slot)
-    return action
-
-
-class RunPreviewCell(qtw.QWidget):
-    plotRequested = QtCore.pyqtSignal(str, str)
-    exportRequested = QtCore.pyqtSignal(str, str)
-
-    def __init__(self, guid, count, parent=None, icon_size=MEASUREMENT_PREVIEW_SIZE):
-        super().__init__(parent)
-        self.guid = guid
-        self.placeholder_count = max(0, int(count or 0))
-        self.icon_size = int(icon_size)
-
-        self.content_layout = qtw.QHBoxLayout()
-        self.content_layout.setContentsMargins(2, 0, 2, 0)
-        self.content_layout.setSpacing(MEASUREMENT_PREVIEW_SPACING)
-        self.content_layout.setAlignment(QtCore.Qt.AlignmentFlag.AlignLeft | QtCore.Qt.AlignmentFlag.AlignVCenter)
-        self.setLayout(self.content_layout)
-        self.setFixedHeight(self.icon_size + 6)
-        self.show_placeholders()
-
-
-    def show_placeholders(self, count=None):
-        self._clear_layout()
-        placeholder_count = self.placeholder_count if count is None else max(0, int(count))
-        for _ in range(placeholder_count):
-            self.content_layout.addWidget(self._placeholder_label())
-        self.content_layout.addStretch()
-
-
-    def show_previews(self, previews):
-        self._clear_layout()
-
-        preview_count = 0
-        for preview in previews or []:
-            image = preview.get("image")
-            if image is None:
-                continue
-
-            label = DraggablePreviewImageLabel(
-                self.guid,
-                preview.get("parameter", ""),
-                preview.get("axes") or [],
-                )
-            label.setObjectName("measurementPreviewImage")
-            label.setFixedSize(self.icon_size, self.icon_size)
-            label.setAlignment(QtCore.Qt.AlignmentFlag.AlignCenter)
-            label.setToolTip(preview.get("title", ""))
-            label.setPixmap(
-                QtGui.QPixmap.fromImage(image).scaled(
-                    self.icon_size,
-                    self.icon_size,
-                    QtCore.Qt.AspectRatioMode.KeepAspectRatio,
-                    QtCore.Qt.TransformationMode.SmoothTransformation,
-                    )
-                )
-            label.plotRequested.connect(self._emit_plot_requested)
-            label.exportRequested.connect(self._emit_export_requested)
-            self.content_layout.addWidget(label)
-            preview_count += 1
-
-        for _ in range(max(0, self.placeholder_count - preview_count)):
-            self.content_layout.addWidget(self._placeholder_label())
-        self.content_layout.addStretch()
-
-
-    def _placeholder_label(self):
-        label = qtw.QLabel()
-        label.setObjectName("measurementPreviewPlaceholder")
-        label.setFixedSize(self.icon_size, self.icon_size)
-        label.setFrameShape(qtw.QFrame.Shape.Box)
-        label.setFrameShadow(qtw.QFrame.Shadow.Plain)
-        label.setLineWidth(1)
-        return label
-
-
-    def _emit_plot_requested(self, parameter):
-        self.plotRequested.emit(self.guid, parameter)
-
-
-    def _emit_export_requested(self, parameter):
-        self.exportRequested.emit(self.guid, parameter)
-
-
-    def _clear_layout(self):
-        while self.content_layout.count():
-            item = self.content_layout.takeAt(0)
-            widget = item.widget()
-            if widget is not None:
-                widget.setParent(None)
-                widget.deleteLater()
-
-
-class EqualsAlignedDelegate(qtw.QStyledItemDelegate):
-    """
-    Paints setpoint counts with equals signs aligned.
-    """
-
-    right_text_alignment = QtCore.Qt.AlignmentFlag.AlignLeft | QtCore.Qt.AlignmentFlag.AlignVCenter
-
-    def paint(self, painter, option, index):
-        text = index.data(QtCore.Qt.ItemDataRole.DisplayRole)
-        if text is None or text == "":
-            super().paint(painter, option, index)
-            return
-
-        opt = qtw.QStyleOptionViewItem(option)
-        self.initStyleOption(opt, index)
-        text = str(text)
-        sections = self._display_sections(text)
-        if sections is None:
-            super().paint(painter, option, index)
-            return
-        left, right = sections
-
-        opt.text = ""
-
-        widget = opt.widget
-        style = widget.style() if widget else qtw.QApplication.style()
-        style.drawControl(qtw.QStyle.ControlElement.CE_ItemViewItem, opt, painter, widget)
-
-        text_rect = style.subElementRect(
-            qtw.QStyle.SubElement.SE_ItemViewItemText,
-            opt,
-            widget
-            ).adjusted(2, 0, -2, 0)
-        metrics = opt.fontMetrics
-        equals_text = " = "
-        equals_width = metrics.horizontalAdvance(equals_text)
-        max_right_width = self._max_right_width(index, metrics)
-        if right is None and max_right_width <= 0:
-            painter.save()
-            painter.setFont(opt.font)
-            painter.setPen(self._text_color(opt))
-            painter.drawText(
-                text_rect,
-                QtCore.Qt.AlignmentFlag.AlignRight | QtCore.Qt.AlignmentFlag.AlignVCenter,
-                metrics.elidedText(left, QtCore.Qt.TextElideMode.ElideLeft, text_rect.width())
-                )
-            painter.restore()
-            return
-
-        equals_left = max(
-            text_rect.left(),
-            text_rect.right() - max_right_width - equals_width
-            )
-
-        left_rect = QtCore.QRect(
-            text_rect.left(),
-            text_rect.top(),
-            max(0, equals_left - text_rect.left()),
-            text_rect.height()
-            )
-        equals_rect = QtCore.QRect(
-            equals_left,
-            text_rect.top(),
-            equals_width,
-            text_rect.height()
-            )
-        right_rect = QtCore.QRect(
-            equals_left + equals_width,
-            text_rect.top(),
-            max(0, text_rect.right() - equals_left - equals_width + 1),
-            text_rect.height()
-            )
-
-        painter.save()
-        painter.setFont(opt.font)
-        painter.setPen(self._text_color(opt))
-        painter.drawText(
-            left_rect,
-            QtCore.Qt.AlignmentFlag.AlignRight | QtCore.Qt.AlignmentFlag.AlignVCenter,
-            metrics.elidedText(left, QtCore.Qt.TextElideMode.ElideLeft, left_rect.width())
-            )
-        if right is not None:
-            painter.drawText(equals_rect, QtCore.Qt.AlignmentFlag.AlignCenter, equals_text)
-            self._draw_right_text(painter, right_rect, right, metrics)
-        painter.restore()
-
-
-    def _display_sections(self, text):
-        if " = " in text:
-            return str(text).split(" = ", 1)
-        if self._is_count_text(text):
-            return str(text), None
-        return None
-
-
-    def _is_count_text(self, text):
-        try:
-            int(str(text).replace(",", ""))
-        except (TypeError, ValueError):
-            return False
-        return True
-
-
-    def _draw_right_text(self, painter, right_rect, right, metrics):
-        painter.drawText(
-            right_rect,
-            self.right_text_alignment,
-            metrics.elidedText(right, QtCore.Qt.TextElideMode.ElideRight, right_rect.width())
-            )
-
-
-    def _max_right_width(self, index, metrics):
-        view = self.parent()
-        if not isinstance(view, qtw.QTreeWidget):
-            return 0
-
-        max_width = 0
-        column = index.column()
-        for row in range(view.topLevelItemCount()):
-            item = view.topLevelItem(row)
-            if item is None:
-                continue
-
-            sections = self._display_sections(item.text(column))
-            if sections is None:
-                continue
-
-            _, right = sections
-            if right is None:
-                continue
-
-            max_width = max(max_width, metrics.horizontalAdvance(right))
-        return max_width
-
-
-    def _text_color(self, option):
-        return option.palette.color(QtGui.QPalette.ColorRole.Text)
 
 
 class RunList(qtw.QTreeWidget):
@@ -816,49 +582,6 @@ class RunList(qtw.QTreeWidget):
                 self.add_plot(target_win, param)
    
     
-#3 classes/methods below are adapted from plottr and as such are not commented by me
-class SortableTreeWidgetItem(qtw.QTreeWidgetItem):
-    """
-    QTreeWidgetItem with an overridden comparator that sorts numerical values
-    as numbers instead of sorting them alphabetically.
-    """
-    def __init__(self, strings):
-        super().__init__(strings)
-        self.run_metadata = {}
-        self._guid = ""
-
-    def __lt__(self, other: qtw.QTreeWidgetItem) -> bool:
-        col = self.treeWidget().sortColumn()
-        value1 = self.data(col, QtCore.Qt.ItemDataRole.UserRole)
-        value2 = other.data(col, QtCore.Qt.ItemDataRole.UserRole)
-        if value1 is not None and value2 is not None:
-            try:
-                return float(value1) < float(value2)
-            except (TypeError, ValueError):
-                pass
-
-        text1 = self.text(col)
-        text2 = other.text(col)
-        try:
-            return float(text1) < float(text2)
-        except ValueError:
-            return text1 < text2    
-    
-    @property
-    def guid(self): # Easier fetching for data
-        return self._guid
-
-
-    def set_guid(self, guid):
-        self._guid = guid
-
-
-    def update_tooltip(self):
-        tooltip = run_tooltip_text(self.run_metadata)
-        for col in range(self.columnCount()):
-            self.setToolTip(col, tooltip)
-
-
 class moreInfo(qtw.QTabWidget):
     
     def __init__(self, *args, preview_size=None):
@@ -1219,281 +942,3 @@ class moreInfo(qtw.QTabWidget):
 
     def _has_value(self, value):
         return value is not None and value != ""
-
-
-class WrappedValueDelegate(qtw.QStyledItemDelegate):
-    WRAP_FLAGS = (
-        QtCore.Qt.AlignmentFlag.AlignLeft
-        | QtCore.Qt.AlignmentFlag.AlignTop
-        | QtCore.Qt.TextFlag.TextWrapAnywhere
-        )
-
-    def paint(self, painter, option, index):
-        opt = qtw.QStyleOptionViewItem(option)
-        self.initStyleOption(opt, index)
-
-        widget = opt.widget
-        style = widget.style() if widget is not None else qtw.QApplication.style()
-        text = opt.text
-        opt.text = ""
-
-        style.drawControl(qtw.QStyle.ControlElement.CE_ItemViewItem, opt, painter, widget)
-
-        text_rect = style.subElementRect(qtw.QStyle.SubElement.SE_ItemViewItemText, opt, widget)
-        text_rect.adjust(0, 2, 0, -2)
-        painter.save()
-        painter.setFont(opt.font)
-        role = (
-            QtGui.QPalette.ColorRole.HighlightedText
-            if opt.state & qtw.QStyle.StateFlag.State_Selected
-            else QtGui.QPalette.ColorRole.Text
-            )
-        painter.setPen(opt.palette.color(role))
-        painter.drawText(text_rect, self.WRAP_FLAGS, text)
-        painter.restore()
-
-
-    def sizeHint(self, option, index):
-        opt = qtw.QStyleOptionViewItem(option)
-        self.initStyleOption(opt, index)
-
-        width = opt.rect.width()
-        if width <= 0 and opt.widget is not None:
-            width = opt.widget.columnWidth(index.column())
-        width = max(24, width - 6)
-
-        metrics = QtGui.QFontMetrics(opt.font)
-        text_rect = metrics.boundingRect(
-            QtCore.QRect(0, 0, width, 100000),
-            self.WRAP_FLAGS,
-            opt.text
-            )
-        base = super().sizeHint(option, index)
-        return QtCore.QSize(base.width(), max(base.height(), text_rect.height() + 6))
-
-
-class infoTree(qtw.QTreeWidget):
-    def __init__(self, expand_all=True, truncate_values=False):
-        super().__init__()
-        self.expand_all = expand_all
-        self.truncate_values = truncate_values
-        self.setHeaderLabels(["Key", "Value"])
-        self.setColumnCount(2)
-        self.setWordWrap(True)
-        self.setTextElideMode(QtCore.Qt.TextElideMode.ElideNone)
-        self.setUniformRowHeights(False)
-        self.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarPolicy.ScrollBarAlwaysOff)
-        self.setItemDelegateForColumn(1, WrappedValueDelegate(self))
-        self.header().setSectionResizeMode(0, qtw.QHeaderView.ResizeMode.ResizeToContents)
-        self.header().setSectionResizeMode(1, qtw.QHeaderView.ResizeMode.Stretch)
-        self.header().setStretchLastSection(True)
-        self.setContextMenuPolicy(QtCore.Qt.ContextMenuPolicy.CustomContextMenu)
-        self.customContextMenuRequested.connect(self.openCopyMenu)
-
-        self.copy_value_action = copy_action(
-            "Copy Value",
-            COPY_CELL_SHORTCUTS,
-            self.copyValue,
-            self
-            )
-        self.copy_selection_action = copy_action(
-            "Copy Selection",
-            COPY_SELECTION_SHORTCUTS,
-            self.copySelection,
-            self
-            )
-        self.addActions([self.copy_value_action, self.copy_selection_action])
-
-
-    def setInfo(self, info):
-        self.clear()
-
-        if not info:
-            self.addTopLevelItem(qtw.QTreeWidgetItem(["No data", ""]))
-            return
-        if not isinstance(info, dict):
-            item = qtw.QTreeWidgetItem(["Value", format_value(info, 180 if self.truncate_values else None)])
-            item.setToolTip(1, format_value(info))
-            self.addTopLevelItem(item)
-            return
-
-        items = dictToTree(info, truncate_values=self.truncate_values)
-        for item in items:
-            self.addTopLevelItem(item)
-            item.setExpanded(True)
-
-        if self.expand_all:
-            self.expandAll()
-        self.header().setSectionResizeMode(0, qtw.QHeaderView.ResizeMode.ResizeToContents)
-        self.header().setSectionResizeMode(1, qtw.QHeaderView.ResizeMode.Stretch)
-        self.doItemsLayout()
-
-
-    def resizeEvent(self, event):
-        super().resizeEvent(event)
-        self.doItemsLayout()
-
-
-    def openCopyMenu(self, pos):
-        item = self.itemAt(pos)
-        if item is None:
-            return
-
-        self.setCurrentItem(item)
-
-        menu = qtw.QMenu(self)
-        menu.addAction(self.copy_value_action)
-
-        copy_row = QtGui.QAction("Copy Row", menu)
-        copy_row.triggered.connect(lambda: copy_to_clipboard(row_text(item)))
-        menu.addAction(copy_row)
-
-        if self.selectedItems():
-            menu.addAction(self.copy_selection_action)
-
-        menu.exec(self.viewport().mapToGlobal(pos))
-
-
-    def copyValue(self):
-        item = self.currentItem()
-        if item is not None:
-            copy_to_clipboard(item.text(1))
-
-
-    def copySelection(self):
-        items = self.selectedItems()
-        if not items and self.currentItem() is not None:
-            items = [self.currentItem()]
-        copy_to_clipboard("\n".join(row_text(item) for item in items))
-
-
-class CopyableTableWidget(qtw.QTableWidget):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.setContextMenuPolicy(QtCore.Qt.ContextMenuPolicy.CustomContextMenu)
-        self.customContextMenuRequested.connect(self.openCopyMenu)
-
-        self.copy_cell_action = copy_action(
-            "Copy Cell",
-            COPY_CELL_SHORTCUTS,
-            self.copyCell,
-            self
-            )
-        self.copy_selection_action = copy_action(
-            "Copy Selection",
-            COPY_SELECTION_SHORTCUTS,
-            self.copySelection,
-            self
-            )
-        self.addActions([self.copy_cell_action, self.copy_selection_action])
-
-
-    def openCopyMenu(self, pos):
-        item = self.itemAt(pos)
-        if item is None:
-            return
-
-        self.setCurrentItem(item)
-
-        menu = qtw.QMenu(self)
-        menu.addAction(self.copy_cell_action)
-        menu.addAction(self.copy_selection_action)
-
-        menu.exec(self.viewport().mapToGlobal(pos))
-
-
-    def copyCell(self):
-        item = self.currentItem()
-        if item is not None:
-            copy_to_clipboard(item.text())
-
-
-    def copySelection(self):
-        ranges = self.selectedRanges()
-        if not ranges and self.currentItem() is not None:
-            self.copyCell()
-            return
-
-        sections = []
-        for selected_range in ranges:
-            rows = []
-            for row in range(selected_range.topRow(), selected_range.bottomRow() + 1):
-                values = []
-                for col in range(selected_range.leftColumn(), selected_range.rightColumn() + 1):
-                    item = self.item(row, col)
-                    values.append(item.text() if item is not None else "")
-                rows.append("\t".join(values))
-            sections.append("\n".join(rows))
-
-        copy_to_clipboard("\n".join(section for section in sections if section))
-
-
-def dictToTree(d : dict, truncate_values=False):
-    items = []
-    for k, v in d.items():
-        if not isinstance(v, dict):
-            item = qtw.QTreeWidgetItem([str(k), format_value(v, 180 if truncate_values else None)])
-            item.setToolTip(1, format_value(v))
-        else:
-            item = qtw.QTreeWidgetItem([k, ''])
-            for child in dictToTree(v, truncate_values=truncate_values):
-                item.addChild(child)
-        items.append(item)
-    return items
-
-
-def snapshot_parameters(snapshot):
-    if not isinstance(snapshot, dict):
-        return {}
-
-    out = {}
-    station = snapshot.get("station", snapshot)
-    parameter_dicts = []
-
-    if isinstance(station, dict):
-        params = station.get("parameters")
-        if isinstance(params, dict):
-            parameter_dicts.append(params)
-
-        instruments = station.get("instruments")
-        if isinstance(instruments, dict):
-            for instrument in instruments.values():
-                if isinstance(instrument, dict) and isinstance(instrument.get("parameters"), dict):
-                    parameter_dicts.append(instrument["parameters"])
-
-    for params in parameter_dicts:
-        for key, details in params.items():
-            if not isinstance(details, dict):
-                continue
-            for name in (key, details.get("name"), details.get("full_name")):
-                if name:
-                    out[str(name)] = details
-
-    return out
-
-
-def format_value(value, max_len=None):
-    if value is None:
-        text = ""
-    elif isinstance(value, float):
-        text = f"{value:.6g}"
-    elif isinstance(value, (list, tuple)):
-        text = ", ".join(format_value(item) for item in value)
-    else:
-        text = str(value)
-
-    text = text.replace("\n", " ")
-    if max_len is not None and len(text) > max_len:
-        return text[:max_len - 3] + "..."
-    return text
-
-
-def row_text(item):
-    return "\t".join(item.text(col) for col in range(item.columnCount()))
-
-
-def copy_to_clipboard(text):
-    clipboard = qtw.QApplication.clipboard()
-    if clipboard is not None:
-        clipboard.setText(text)
-
